Fix image type

+++ stpt_pipeline/stpt_bead_registration.py
@@ -61,7 +61,7 @@ def _get_labels(im, pedestal, im_std):
     im_temp = labels > 0
     distance = ndi.distance_transform_edt(im_temp)
     local_maxi = peak_local_max(
-        distance, indices=False, footprint=np.ones((3, 3)), labels=im_temp
+        distance, indices=False, footprint=np.ones((3, 3)), labels=im_temp.astype('int')
     )
     marks = ndi.label(local_maxi)[0]
     labels = watershed(-distance, marks, mask=im_temp)
